Update-2021.01.05 - Update Metadynamics

+++ PyPol/gromacs.py
@@ -2862,8 +2862,8 @@ COMMITTOR ...
                       "<<< 0 &> /dev/null".format(self, file_ext, traj_file, time))
             os.system("{0._gromacs} trjconv -f TMP_PYPOL{1} -o plumed_{2}.xtc -b {4} -e {3} -s {0._name}.tpr "
                       "<<< 0 &> /dev/null".format(self, file_ext, file_name, time, time - timeinterval))
-
-            os.rename(crystal._path + file_name + ".gro", crystal._path + file_name + "_old.gro")
+            if os.path.exists(crystal._path + file_name + ".gro"):
+                os.rename(crystal._path + file_name + ".gro", crystal._path + file_name + "_old.gro")
             os.system("{0._gromacs} trjconv -f TMP_PYPOL{1} -o {2}.gro -b {3} -dump {4} -s {0._name}.tpr "
                       "<<< 0 &> /dev/null".format(self, file_ext, file_name, time - 100, time))
